Use non-blocking / non-signal based timeout
Parallel::ForkManager's start() method is a non-blocking loop that waits until
there is available capacity to start another child process, watching until other
child processes have finished. We can add some logic that it executes during its
wait loop to see if we've passed the timeout duration. If so, we can have it
kill the current running children and make sure we no longer try to start any
more children thereafter.

When we kill the children, this will let it exit its loop in start() so that it
is no longer "stuck" in that loop as killing the children frees up capacity.

However, we still need to see if, after all children were successfully
dispatched, that those final children may also be exceeding our global timeout
as well and kill them off as needed. This is also done in a non-blocking loop as
we try to wait for all finished children.